refactor: Rename folder to match naming conventions
