+++ src/app/(after-login)/my-info/login-info/page.tsx
@@ -9,11 +9,11 @@ import GoogleIcon from '@/assets/provider/Google.svg';
 import KakaoIcon from '@/assets/provider/Kakao.svg';
 import Button from '@/components/common/button';
 import Dialog from '@/components/common/dialog';
-import { SizedBox } from '@/components/common/sizedbox/SizedBox';
 import { TopNavBar } from '@/components/common/top-navbar';
 import Layout from '@/components/layout';
 
 import styles from './page.module.scss';
+import { SizedBox } from '@/components/common/sizedbox';
 
 const BLOCK = 'page';
