+++ src/components/common/menubox/index.tsx
@@ -3,7 +3,7 @@ import {forwardRef} from 'react';
 import {Menu, MenuButton, MenuItems} from '@headlessui/react';
 import classNames from 'classnames/bind';
 
-import {ListItem, ListItemProps} from './ListItem';
+import {ListItem, ListItemProps} from './list-item';
 import styles from './index.module.scss';
 
 const cx = classNames.bind(styles);
